+++ app/Http/Requests/UpdateProductRequest.php
@@ -13,7 +13,7 @@ class UpdateProductRequest extends FormRequest
         return [
             'name' => ['sometimes', 'required', 'string', 'max:255'],
             'description' => ['nullable', 'string'],
-            'price' => ['sometimes', 'required', 'numeric', 'min:0'],
+            'price' => ['sometimes', 'required', 'integer', 'min:0'],
             'stock' => ['sometimes', 'required', 'integer', 'min:0'],
             'cover_image' => ['nullable', 'image', 'mimes:jpeg,png,jpg,gif,svg', 'max:5120'],
         ];
